Move web module in rust-code-analysis-cli workspace
