+++ Cargo.toml
@@ -10,12 +10,8 @@ cc = "^1.0"
 phf_codegen = "^0.8"
 
 [dependencies]
-actix-rt = "^1.0"
-actix-web = "^2.0"
 aho-corasick = "^0.7"
-bytes = "^0.5"
 enum-iterator = "^0.6"
-futures = "^0.3"
 fxhash = "0.2"
 json = "^0.12"
 lazy_static = "^1.3"
@@ -23,7 +19,7 @@ num-format = "^0.4"
 petgraph = "^0.5"
 phf = { version = "^0.8", features = ["macros"] }
 regex = "^1.1"
-serde = "^1.0"
+serde = { version = "^1.0", features = ["derive"] }
 serde_cbor = "^0.11"
 serde_json = "^1.0"
 termcolor = "^1.0"
