+++ rust-code-analysis-cli/Cargo.toml
@@ -5,10 +5,18 @@ authors = ["Calixte Denizet <cdenizet@mozilla.com>"]
 edition = "2018"
 
 [dependencies]
+actix-rt = "^1.0"
+actix-web = "^2.0"
+bytes = "^0.5"
 clap = "^2.33"
 crossbeam = "^0.7"
+futures = "^0.3"
 globset = "^0.4"
 num_cpus = "^1.13"
 rust-code-analysis = { path = ".." }
+serde = "^1.0"
 serde_json = "^1.0"
 walkdir = "^2.2"
+
+[dev-dependencies]
+pretty_assertions = "^0.6"
