+++ rust-code-analysis-cli/src/web/mod.rs
@@ -1,5 +1,3 @@
-pub mod alterator;
-pub mod ast;
 pub mod comment;
 pub mod function;
 pub mod metrics;
